advance_ml_webapp.py
Best and advance webapp for all the users in this webapp we can use all the library in it. first we add some slightbar in it to change the data according to the user need it can also change it and also change the plots in this webapp. Second, we can create create a webapp and addings plots in it. Third, we can also add some of the machine learning models in it also add some prediction in it by using iris data set